refactor trimming
