+++ src/parser.rs
@@ -301,3 +301,5 @@ impl LargeUtf8Parser {
         }
     }
 }
+
+
